support dry-run and improve debug logging

+++ himlarcli/keystone.py
@@ -271,16 +271,26 @@ class Keystone(Client):
             return None
         if not region:
             region = self.find_regions()
+
         # Delete DNS zones
-        self.__delete_zones(project)
+        deleted_zones = self.__delete_zones(project)
+        for zone in deleted_zones:
+            self.debug_log('DELETED zone: {}'. format(zone))
+
         # Delete instances
         self.__delete_instances(project, region)
+
         # Revoke shared image memberships
-        self.__revoke_image_shares(project, region)
+        revoked_images = self.__revoke_image_shares(project, region)
+        for image in revoked_images:
+            self.debug_log('REVOKED access to image: {} ({})'. format(image, region))
+
         # Delete images
         self.__delete_images(project, region)
+
         # Delete security groups
         self.__delete_security_groups(project, region)
+
         # Delete volume
         self.__delete_volumes(project, region)
 
@@ -822,10 +832,22 @@ class Keystone(Client):
     def __delete_zones(self, project):
         """ Use designateclient to delete all zones belonging to a project
             version: 2021-06 """
+
+        # Initiate Designate object
         dc = self._get_client(Designate)
+
+        # Get a list of zones belonging to project
         zones = dc.list_project_zones(project.id)
+
+        # Delete the zones, collecting the names of the deleted zones
+        deleted_zones = []
         for zone in zones:
-            dc.delete_zone(zone['id'])
+            if not self.dry_run:
+                dc.delete_zone(zone['id'])
+            deleted_zones.append(zone['name'])
+
+        # Return a list of deleted zone names
+        return deleted_zones
 
     def __revoke_image_shares(self, project, region):
         """ Use glanceclient to revoke membership for any shared images that
@@ -843,13 +865,17 @@ class Keystone(Client):
             images = gc.get_images(filters=filters)
 
             # Remove image member status for images shared to this project
+            revoked_images = []
             for image in images:
                 members = gc.get_image_access(image_id=image.id)
                 for member in members:
                     if member.member_id == project.id:
-                        gc.set_image_access(image_id=image.id, project_id=project.id, action='revoke')
-                        self.debug_log('REVOKE access to image {} for project {}'.
-                                       format(image.name, project.name))
+                        if not self.dry_run:
+                            gc.set_image_access(image_id=image.id, project_id=project.id, action='revoke')
+                        revoked_images.append("%s (%s)" % (image.name, region))
+
+            # Return a list of revoked images
+            return revoked_images
 
     def __delete_images(self, project, region):
         """ Use glanceclient to delete all private images for a project in
